refactor authorizations page + modal
