+++ deploy-web/src/app/settings/authorizations/AllowanceIssuedRow.tsx
@@ -1,14 +1,14 @@
+"use client";
 import React, { ReactNode } from "react";
-import { IconButton, TableCell } from "@mui/material";
-import { CustomTableRow } from "../shared/CustomTable";
-import { Address } from "../shared/Address";
 import { FormattedTime } from "react-intl";
-import EditIcon from "@mui/icons-material/Edit";
-import DeleteIcon from "@mui/icons-material/Delete";
 import { AllowanceType } from "@src/types/grant";
-import { AKTAmount } from "../shared/AKTAmount";
 import { coinToUDenom } from "@src/utils/priceUtils";
 import { getAllowanceTitleByType } from "@src/utils/grants";
+import { TableCell, TableRow } from "@src/components/ui/table";
+import { Address } from "@src/components/shared/Address";
+import { AKTAmount } from "@src/components/shared/AKTAmount";
+import { Button } from "@src/components/ui/button";
+import { Bin, Edit } from "iconoir-react";
 
 type Props = {
   allowance: AllowanceType;
@@ -18,10 +18,8 @@ type Props = {
 };
 
 export const AllowanceIssuedRow: React.FunctionComponent<Props> = ({ allowance, onEditAllowance, setDeletingAllowance }) => {
-  // const denomData = useDenomData(grant.authorization.spend_limit.denom);
-
   return (
-    <CustomTableRow>
+    <TableRow>
       <TableCell>{getAllowanceTitleByType(allowance)}</TableCell>
       <TableCell>
         <Address address={allowance.grantee} isCopyable />
@@ -33,13 +31,13 @@ export const AllowanceIssuedRow: React.FunctionComponent<Props> = ({ allowance, 
         <FormattedTime year="numeric" month={"numeric"} day={"numeric"} value={allowance.allowance.expiration} />
       </TableCell>
       <TableCell align="right">
-        <IconButton onClick={() => onEditAllowance(allowance)}>
-          <EditIcon />
-        </IconButton>
-        <IconButton onClick={() => setDeletingAllowance(allowance)}>
-          <DeleteIcon />
-        </IconButton>
+        <Button onClick={() => onEditAllowance(allowance)} variant="ghost" size="icon">
+          <Edit />
+        </Button>
+        <Button onClick={() => setDeletingAllowance(allowance)} variant="ghost" size="icon" className="ml-2">
+          <Bin />
+        </Button>
       </TableCell>
-    </CustomTableRow>
+    </TableRow>
   );
 };
