+++ deploy-web/src/components/shared/Fieldset.tsx
@@ -4,12 +4,13 @@ import { Card, CardContent } from "../ui/card";
 
 type Props = {
   label: string;
+  className?: string;
   children: React.ReactNode;
 };
 
-export const Fieldset: React.FunctionComponent<Props> = ({ label, children }) => {
+export const Fieldset: React.FunctionComponent<Props> = ({ label, className = "", children }) => {
   return (
-    <Card>
+    <Card className={className}>
       <CardContent className="relative rounded-sm">
         <div className="p-2 text-muted-foreground">
           <p>{label}</p>
